+++ src/main/java/kr/hhplus/be/server/interfaces/product/dto/ProductResponse.java
@@ -37,7 +37,7 @@ public record ProductResponse() {
             Long optionId,
             String optionValue,
             Long price,
-            Long stock
+            Integer stock
     ) {
         public static Option from(ProductResult.Option option) {
             return Option.builder()
